add to cart and navigate to cart

+++ src/features/cart/cartSlice.js
@@ -4,7 +4,7 @@ import { createSlice } from '@reduxjs/toolkit'
 
 if (!localStorage.getItem("giacat")) {
   localStorage.setItem("giacat", JSON.stringify([
-    { id: '1', name: 'Máy chạy bộ Impulse PT300', price: 45000000, image: "https://firebasestorage.googleapis.com/v0/b/fitness-mall---hcmut.appspot.com/o/item-images%2FPT300_may-chay-bo-impulse.jpg?alt=media&token=ae7a34bd-09ff-4c65-8de6-fd54dcdb7e4e", quantity: 1, isChosen: false },
+    // { id: '1', name: 'Máy chạy bộ Impulse PT300', price: 45000000, image: "https://firebasestorage.googleapis.com/v0/b/fitness-mall---hcmut.appspot.com/o/item-images%2FPT300_may-chay-bo-impulse.jpg?alt=media&token=ae7a34bd-09ff-4c65-8de6-fd54dcdb7e4e", quantity: 1, isChosen: false },
     // { id: '2', name: 'Tạ tay Brosman', price: 65000, image: "https://firebasestorage.googleapis.com/v0/b/fitness-mall---hcmut.appspot.com/o/item-images%2FBM5_ta-tay-brosman.jpg?alt=media&token=153c9fc5-8f7f-43a1-bff5-e10a41e692c5", quantity: 1, isChosen: false },
   ]))
 }
@@ -51,10 +51,13 @@ const cartSlice = createSlice({
     },
     addItem: {
       reducer(state, action) {
-        state.push(action.payload)
-        localStorage.setItem("giacat", JSON.stringify(state))
+        const existingItem = state.find(item => item.id === action.payload.id)
+        if (!existingItem) {
+          state.push(action.payload)
+          localStorage.setItem("giacat", JSON.stringify(state))
+        }
       },
-      prepare(id, name, price, image) {
+      prepare(id, name, price, image, quantity) {
         return {
           payload: {
             id,
